kj-http: Support suspending and resuming HTTP requests
To accomplish this, we add a suspendable HttpServer::listenHttpCleanDrain() overload. The new overload takes three parameters: the AsyncIoStream& connection, a SuspendableHttpServiceFactory which can construct custom HttpService objects or suspend the request based on the incoming request headers, and a Maybe<SuspendedRequest> which provides the ability to resume requests.

These changes turned out a bit hackier than I'd like.